Use ComputerPlayerII in place of the old ComputerPlayer
Rename ComputerPlayerII to ComputerPlayer, and rename ComputerPlayer
to OldComputerPlayer.
